Import live.2013.09.08.tar.gz
2013.09.07
- Improved the bugfix that was in the previous release (because that fix still had a deleted object being accessed).

2013.09.08
- Made a minor change to the way that "openRTSP" handles the "-Q" option (to address an issue that someone reported).

+++ testProgs/playCommon.cpp
@@ -1046,8 +1046,7 @@ static void scheduleNextQOSMeasurement() {
   struct timeval timeNow;
   gettimeofday(&timeNow, NULL);
   unsigned timeNowUSecs = timeNow.tv_sec*1000000 + timeNow.tv_usec;
-  unsigned usecsToDelay = nextQOSMeasurementUSecs - timeNowUSecs;
-     // Note: This works even when nextQOSMeasurementUSecs wraps around
+  int64_t usecsToDelay = nextQOSMeasurementUSecs - timeNowUSecs;
 
   qosMeasurementTimerTask = env->taskScheduler().scheduleDelayedTask(
      usecsToDelay, (TaskFunc*)periodicQOSMeasurement, (void*)NULL);
